Se agrego el doble chequeo de pass

+++ frontend/js/perfil.js
@@ -589,12 +589,24 @@ async function eliminarUsuario() {
   const confirmar = confirm(
     "⚠️¿Seguro que querés eliminar tu cuenta?"
   );
+  const contrasenia = document.getElementById("contraElim");
+  const confirmarContra = document.getElementById("contraConfirm");
+  const id = localStorage.getItem("idUsuario");
 
   if (!confirmar) return;
 
+  if (!contrasenia.value || !confirmarContra.value) {
+    alert("Todos los campos son obligatorios");
+    return;
+  }
+
+  if(contrasenia.value !== confirmarContra.value){
+      alert("Los datos no coinciden");
+      return;
+  }
+
   try {
-    const contrasenia = document.getElementById("contraElim");
-    const id = localStorage.getItem("idUsuario");
+    
     console.log(id);
     const response = await fetch("http://localhost:3000/usuarios", {
       method: "DELETE",
